requirements.txt
autopep8==2.0.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
python-dotenv==0.20.0
tomli==2.0.1
Werkzeug==2.2.2
